+++ packages/react/src/components/Card/Card.jsx
@@ -215,6 +215,7 @@ export const defaultProps = {
     extra: false,
   },
   renderExpandIcon: undefined,
+  renderDateDropdownInPortal: true,
   rowHeight: ROW_HEIGHT,
   breakpoint: DASHBOARD_SIZES.LARGE,
   cardDimensions: CARD_DIMENSIONS,
@@ -299,6 +300,7 @@ const Card = (props) => {
     onCardAction,
     availableActions,
     renderExpandIcon,
+    renderDateDropdownInPortal,
     breakpoint,
     i18n,
     style,
@@ -447,6 +449,7 @@ const Card = (props) => {
       width={cardSize.width}
       availableActions={mergedAvailableActions}
       renderExpandIcon={renderExpandIcon}
+      renderDateDropdownInPortal={renderDateDropdownInPortal}
       i18n={strings}
       isEditable={isEditable}
       isExpanded={isExpanded}