+++ packages/react/src/components/DateTimePicker/DateTimePickerV2WithoutTimeSpinner.jsx
@@ -170,6 +170,10 @@ const propTypes = {
   hasIconOnly: PropTypes.bool,
   /** Allow repositioning the flyout menu */
   menuOffset: PropTypes.shape({ left: PropTypes.number, top: PropTypes.number }),
+  /** If set to true it will render outside of the current DOM in a portal, otherwise render as a child */
+  renderInPortal: PropTypes.bool,
+  /** Auto reposition if flyout menu offscreen */
+  useAutoPositioning: PropTypes.bool,
   style: PropTypes.objectOf(PropTypes.string),
 };
 
@@ -254,6 +258,8 @@ const defaultProps = {
   id: undefined,
   hasIconOnly: false,
   menuOffset: undefined,
+  renderInPortal: true,
+  useAutoPositioning: false,
   style: {},
 };
 
@@ -279,6 +285,8 @@ const DateTimePicker = ({
   id = uuidv4(),
   hasIconOnly,
   menuOffset,
+  renderInPortal,
+  useAutoPositioning,
   style,
   ...others
 }) => {
@@ -681,7 +689,7 @@ const DateTimePicker = ({
     ? 0
     : hasIconOnly
     ? -15
-    : 274;
+    : 288;
   const menuOffsetTop = menuOffset?.top ? menuOffset.top : 0;
 
   return (
@@ -788,7 +796,8 @@ const DateTimePicker = ({
             direction={FlyoutMenuDirection.BottomEnd}
             customFooter={CustomFooter}
             tooltipFocusTrap={false}
-            renderInPortal
+            renderInPortal={renderInPortal}
+            useAutoPositioning={useAutoPositioning}
             tooltipClassName={classnames(`${iotPrefix}--date-time-picker--tooltip`, {
               [`${iotPrefix}--date-time-picker--tooltip--icon`]: hasIconOnly,
             })}