+++ packages/react/src/constants/CardPropTypes.js
@@ -973,4 +973,6 @@ export const CardPropTypes = {
   overrides: PropTypes.shape({
     errorMessage: OverridePropTypes,
   }),
+  /** If set to true it will render outside of the current DOM in a portal, otherwise render as a child */
+  renderDateDropdownInPortal: PropTypes.bool,
 };