+++ packages/monorepo-utils/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.151.1-next.0](https://github.com/carbon-design-system/carbon-addons-iot-react/compare/v2.151.0...v2.151.1-next.0) (2022-11-01)
+
+**Note:** Version bump only for package ai-apps-monorepo-utils
+
+
+
+
+
+# [2.151.0](https://github.com/carbon-design-system/carbon-addons-iot-react/compare/v2.151.0-next.47...v2.151.0) (2022-11-01)
+
+**Note:** Version bump only for package ai-apps-monorepo-utils
+
+
+
+
+
 ## [2.150.1-next.0](https://github.com/carbon-design-system/carbon-addons-iot-react/compare/v2.150.0...v2.150.1-next.0) (2022-06-06)
 
 **Note:** Version bump only for package ai-apps-monorepo-utils